编写schedule模块，引入提交延迟文章功能，修改审核为获取到延迟队列才进行审核

+++ yz-leadnews-model/src/main/java/com/leadnews/model/wemedia/dtos/WmNewsDTO.java
@@ -19,7 +19,7 @@ public class WmNewsDTO {
     /**
      * 频道id
      */
-    private Integer channelId;
+    private Long channelId;
     /**
      * 标签
      */
